feat: implementar autenticación con Clerk
