+++ app/(auth)/sign-in/[[...sign-in]]/page.tsx
@@ -1,0 +1,10 @@
+import { SignIn } from "@clerk/nextjs";
+
+export default function Page() {
+  return (
+    <div className="h-screen flex justify-center items-center">
+      <SignIn />
+    </div>
+      
+  );
+}
