+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@clerk/nextjs": "^6.3.4",
     "@eslint/config-array": "^0.19.0",
     "@eslint/object-schema": "^2.1.4",
     "glob": "^11.0.0",
